productivy calculation by dates done

+++ src/main/kotlin/com/aegro/domain/farm/model/Plot.kt
@@ -3,7 +3,7 @@ package com.aegro.domain.farm.model
 data class Plot (
         val id: String?,
         val description: String?,
-        val area: Long,
+        val area: Float,
         val harvests: List<Harvest>?
 ){
 }
